Added option to SSH connection to prevent storing entries into `known_hosts`
Because the tunneling port is random, every port will be added to `known_hosts` for the same signature. This will prevent storing the signature of the connection.

It is still a known issue that it prompts to add the connection, but this workaround prevents it from being permanently added to `known_hosts`.